She funky on my moment til I peter (333 Project Errors)
https://www.youtube.com/watch?v=bL3baWc-u0Q

+++ src/main/java/org/violetmoon/quark/content/world/module/SpiralSpiresModule.java
@@ -104,8 +104,8 @@ public class SpiralSpiresModule extends ZetaModule {
 			return;
 
 		Entity entity = event.getEntity();
-		Level world = entity.level;
-		BlockPos pos = new BlockPos(event.getTargetX(), event.getTargetY(), event.getTargetZ());
+		Level world = entity.level();
+		BlockPos pos = new BlockPos((int) event.getTargetX(), (int) event.getTargetY(), (int) event.getTargetZ());
 
 		List<BlockPos> myalite = getAdjacentMyalite(null, world, pos, null);
 		if (myalite == null || myalite.isEmpty()) {
